naming수정, 파일합병
